test(storage): move planner tests to tests dir
