+++ crates/floe-core/tests/io_tests/storage/mod.rs
@@ -1,5 +1,6 @@
 pub mod inputs;
 pub mod local;
 pub mod paths;
+pub mod planner;
 pub mod s3;
 pub mod target;
